Create README.txt
FGModel40 is a 3D shape dataset built from ModelNet40 and consists of 2016 shapes (in voxel format with resolution 100*100*100 distributed to 40 categories in an uneven distribution, where shapes number in each category ranges from 18 to 113 unevenly. Shapes in this dataset contain interior structure information and can be used for evaluation to 3D shape retrieval or 3D shape recognition with few examples.